Huge cleanup for events and input.
