+++ slider/src/brain/app.h
@@ -5,8 +5,7 @@
 #include "appConfig.h"
 #include "menu.h"
 #include "src/input/input.h"
-#include "src/input/inputDispatcher.h"
-#include "src/input/inputReader.h"
+#include "src/input/event.h"
 #include "src/output/display.h"
 #include "src/output/displayBuffer.h"
 
@@ -30,7 +29,7 @@ namespace Slider
         std::unique_ptr<Display> m_Display;
         std::unique_ptr<Input::InputReader> m_InputReader;
         DisplayBuffer m_DisplayBuffer;
-        InputDispatcher m_InputDispatcher;
+        EventDispatcher m_InputDispatcher;
     };
 }
